Split CbvFlowsController into multiple controllers
Multiple controllers will be easier to test, reason about, and
simultaneously edit as we continue development. This commit splits up
the controllers as such:

* All CBV flow pages still have URLs under `/cbv/`
* The controllers are nested in the `app/controllers/cbv` directory, and
  are within a new `Cbv` Ruby module accordingly. E.g. `/cbv/entry` is
  in `app/controllers/cbv/entries_controller.rb` and the class name is
  `Cbv::EntriesController`.
* Common functionality is extracted into a `Cbv::BaseController` which
  all pages inherit from.

This refactor also changes a couple other things:
1. The invitation URL is now not in the `/cbv/` URL path. It is now
   `/invitations/new?secret=[secret]`. I've added a redirect.
2. Moves I18n keys to relative keys where possible

[Finishes FFS-807](https://jiraent.cms.gov/browse/FFS-807)
