+++ app/app/controllers/cbv/entries_controller.rb
@@ -1,0 +1,4 @@
+class Cbv::EntriesController < Cbv::BaseController
+  def show
+  end
+end
