+++ app/app/views/cbv/employer_searches/show.html.erb
@@ -1,20 +1,20 @@
 <h2>
-  <%= t('cbv_flows.employer_search.header') %>
+  <%= t('.header') %>
 </h2>
 
-<div data-controller="cbv-flows">
+<div data-controller="cbv-employer-search">
   <h3 class="site-preview-heading">
-    <%= t('cbv_flows.employer_search.subheader') %>
+    <%= t('.subheader') %>
   </h3>
   <%= form_with url: cbv_flow_employer_search_path, method: :get, class: 'usa-search usa-search--big', html: { role: 'search' }, data: { turbo_frame: 'employers', turbo_action: 'advance' } do |f| %>
     <%= f.label :query, "Search for your employer", class: "usa-sr-only" %>
-    <%= f.text_field :query, value: @query, class: "usa-input", type: 'search', data: { "cbv-flows-target": "searchTerms" } %>
+    <%= f.text_field :query, value: @query, class: "usa-input", type: 'search', data: { "cbv-employer-search-target": "searchTerms" } %>
     <button
       class="usa-button"
       type="submit"
     >
       <span class="usa-search__submit-text">
-        <%= t('cbv_flows.employer_search.search') %>
+        <%= t('.search') %>
       </span>
       <%= image_tag "@uswds/uswds/dist/img/usa-icons-bg/search--white.svg", class: "usa-search__submit-icon", alt: "Search" %>
     </button>
@@ -22,8 +22,8 @@
 
   <%= render partial: 'employer', locals: { employer: @employers } %>
 
-  <%= form_with url: next_path, method: :get, class: 'display-none', data: { 'cbv-flows-target': "form" } do |f| %>
-    <input type="hidden" name="user[account_id]" data-cbv-flows-target="userAccountId" />
+  <%= form_with url: next_path, method: :get, class: 'display-none', data: { 'cbv-employer-search-target': "form" } do |f| %>
+    <input type="hidden" name="user[account_id]" data-cbv-employer-search-target="userAccountId" />
   <% end %>
 
   <a
@@ -31,7 +31,7 @@
     class="display-none"
     aria-controls="data-fetching-modal"
     data-open-modal
-    data-cbv-flows-target="modal"
+    data-cbv-employer-search-target="modal"
   ></a>
   <div class="margin-y-3">
     <div
@@ -44,11 +44,11 @@
       <div class="usa-modal__content">
         <div class="usa-modal__main">
           <h2 class="usa-modal__heading" id="modal-3-heading">
-            <%= t('cbv_flows.employer_search.fetching_payroll') %>
+            <%= t('.fetching_payroll') %>
           </h2>
           <div class="usa-prose">
             <p id="modal-3-description">
-              <%= t('cbv_flows.employer_search.fetching_payroll_description') %>
+              <%= t('.fetching_payroll_description') %>
             </p>
           </div>
           <div class="lds-ripple" style="left: 30%;"><div></div><div></div></div>
