+++ app/app/views/cbv/shares/show.html.erb
