+++ app/app/views/cbv/summaries/show.html.erb
