+++ app/app/views/cbv/summaries/show.pdf.erb
@@ -1,0 +1,38 @@
+<p>
+  <%= t('.case_number', case_number: @cbv_flow.case_number) %>
+</p>
+<p>
+  <%= t('.document_timestamp', timestamp: format_view_datetime(Time.now.to_s)) %></p>
+<h2>
+  <%= t('.header') %>
+</h2>
+
+<% if @payments.any? %>
+  <p>
+    <%= t('.description') %>
+  </p>
+  <h3 class="site-preview-heading"><%= t('.total_payments') %> <%= number_to_currency(@payments.reduce(0) { |sum, payment| sum + payment[:amount] }) %></h3>
+
+  <div class="usa-accordion usa-accordion--multiselectable margin-bottom-4" data-allow-multiple>
+    <% @payments.each_with_index do |payment, index| %>
+      <div class="usa-accordion__heading">
+        <%= payment[:employer] %>: <%= t('.payment_of', amount: number_to_currency(payment[:amount])) %>
+      </div>
+      <div id="<%= index %>" class="usa-accordion__content usa-prose">
+        <p>
+          <%= t('.time_range', hours: payment[:hours], start: format_date(payment[:start]), end: format_date(payment[:end])) %>
+        </p>
+      </div>
+    <% end %>
+  </div>
+<% else %>
+  <h3 class="site-preview-heading">
+    <%= t('.none_found') %>
+  </h3>
+<% end %>
+
+<% if @cbv_flow.additional_information.present? %>
+  <h2><%= t('.additional_information_title') %></h2>
+
+  <p><%= @cbv_flow.additional_information %></p>
+<% end %>
