--- app/app/views/cbv_flows/entry.html.erb
@@ -1,19 +1,0 @@
-<h1><%= t('cbv_flows.entry.header') %></h1>
-
-<p><%= t('cbv_flows.entry.subheader') %></p>
-
-<p><%= t('cbv_flows.entry.case_number') %>: <strong><%= @cbv_flow.case_number %></strong></p>
-
-<div class="usa-alert usa-alert--info margin-bottom-3">
-  <div class="usa-alert__body">
-    <p class="usa-alert__text">
-      <%= t('cbv_flows.entry.security_message') %>
-    </p>
-  </div>
-</div>
-
-<%= link_to next_path do %>
-  <button class="usa-button usa-button--outline" type="button">
-    <%= t('cbv_flows.entry.get_started') %>
-  </button>
-<% end %>
