+++ app/app/views/pages/home.html.erb
@@ -4,5 +4,5 @@
 <p><%= t('shared.verify.subheader') %></p>
 
 <%= link_to cbv_flow_entry_path do %>
-  <button class="usa-button usa-button--outline" type="button"><%= t('cbv_flows.manual_flow_start') %></button>
+  <button class="usa-button usa-button--outline" type="button"><%= t('cbv.manual_flow_start') %></button>
 <% end %>
