+++ app/config/locales/en.yml
@@ -9,6 +9,49 @@ en:
       greeting: Hello,
       subject: Verify your income for your benefit application
       thank_you: Thank you!
+  cbv:
+    employer_searches:
+      show:
+        employer_not_listed: Employer not listed?
+        fetching_payroll: We are fetching your employer's payment information.
+        fetching_payroll_description: This may take a few minutes. Please do not close this window.
+        header: Get payment info from your employer.
+        results: Results
+        search: Search
+        search_by_payroll_provider: Search for payroll provider
+        select: Select
+        subheader: Search for an employer
+    entries:
+      show:
+        case_number: Case Number
+        get_started: Get Started
+        header: Let's verify your payment information
+        security_message: Your information is secure. We will not share or keep your log in information. You will be able to preview and approve everything that is shared with your caseworker.
+        subheader: We'll connect to your payment records to get proof of payments to your caseworker and get benefits faster.
+    error_invalid_token: The invitation link used is not valid. Double check the link and try again. If you continue experiencing issues, contact your caseworker.
+    manual_flow_start: Start Flow Manually
+    shares:
+      show:
+        body: Thank you for completing income verification. Click the button below to preview your results and share it with your caseworker.
+        download_pdf: Download PDF
+        header: Share your income with your caseworker
+        share_with_caseworker: Share with Case Worker
+    summaries:
+      show:
+        additional_information_title: Is there anything else you'd like your caseworker to know about your income?
+        case_number: 'Case number: %{case_number}'
+        description: Below is payment information from your employer for the past 90 days.
+        document_timestamp: 'Document origination: %{timestamp}'
+        form:
+          additional_info:
+            label: Is there anything else you'd like your caseworker to know about your income?
+          subheader: Confirm your income amount
+          submit: Continue
+        header: Review payment information from your employer
+        none_found: No payments found. Please refresh the page after a few minutes.
+        payment_of: Payment of %{amount}
+        time_range: "%{hours} hours from %{start} to %{end}"
+        total_payments: Total Payments
   cbv_flow_invitations:
     create:
       invite_failed: 'Error sending invitation to %{email_address}: %{error_message}.'
@@ -21,45 +64,6 @@ en:
         email_address: Email Address
         submit: Send Invitation
       header: Invite someone to complete the CBV flow
-  cbv_flows:
-    employer_search:
-      employer_not_listed: Employer not listed?
-      fetching_payroll: We are fetching your employer's payment information.
-      fetching_payroll_description: This may take a few minutes. Please do not close this window.
-      header: Get payment info from your employer.
-      results: Results
-      search: Search
-      search_by_payroll_provider: Search for payroll provider
-      select: Select
-      subheader: Search for an employer
-    entry:
-      case_number: Case Number
-      error_invalid_token: The invitation link used is not valid. Double check the link and try again. If you continue experiencing issues, contact your caseworker.
-      get_started: Get Started
-      header: Let's verify your payment information
-      security_message: Your information is secure. We will not share or keep your log in information. You will be able to preview and approve everything that is shared with your caseworker.
-      subheader: We'll connect to your payment records to get proof of payments to your caseworker and get benefits faster.
-    manual_flow_start: Start Flow Manually
-    share:
-      body: Thank you for completing income verification. Click the button below to preview your results and share it with your caseworker.
-      download_pdf: Download PDF
-      header: Share your income with your caseworker
-      share_with_caseworker: Share with Case Worker
-    summary:
-      additional_information_title: Is there anything else you'd like your caseworker to know about your income?
-      case_number: 'Case number: %{case_number}'
-      description: Below is payment information from your employer for the past 90 days.
-      document_timestamp: 'Document origination: %{timestamp}'
-      form:
-        additional_info:
-          label: Is there anything else you'd like your caseworker to know about your income?
-        subheader: Confirm your income amount
-        submit: Continue
-      header: Review payment information from your employer
-      none_found: No payments found. Please refresh the page after a few minutes.
-      payment_of: Payment of %{amount}
-      time_range: "%{hours} hours from %{start} to %{end}"
-      total_payments: Total Payments
   shared:
     banner:
       gov_description_html: A <strong>.gov</strong> website belongs to an official government organization in the United States.
